Updated! Added features; Inventory System+ Wearing Clothes+ Interactable Shop Owner NPC+ Clothes Animation System+ Added Build Vol 0.1 to GitHub+
Various updates and improvements have been made. Compared to the old version, it now has a clearer gameplay.

And by the way, you can now shop and try on your dresses with $4000 in your pocket!

You can't sell clothes in the store right now, but don't worry, it will be added soon!